+++ core/image/image.php
@@ -433,7 +433,7 @@ class image
 	 * Generate link to image
 	 *
 	 * @param    string $content what's in the link: image_name, thumbnail, fake_thumbnail, medium or lastimage_icon
-	 * @param    string $mode where does the link lead to: highslide, lytebox, lytebox_slide_show, image_page, image, none
+	 * @param    string $mode where does the link lead to: image_page, image, next, none, or an event-provided mode
 	 * @param    int $image_id
 	 * @param    string $image_name
 	 * @param    int $album_id